PROJECT HEAD/lawglance_project/lawglance_project - Copy/requirements.txt
# Core dependencies
langchain==0.1.3
langchain-core==0.1.16  # Updated to be compatible with langchain 0.1.3

langchain-community==0.0.14
langchain-openai==0.0.5

# AI/ML dependencies
transformers>=4.43.0
spacy==3.7.2
nltk>=3.9
textstat==0.7.3
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.24.3
huggingface-hub>=0.23.2
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
sentence-transformers==2.2.2

# Document processing
python-docx==1.0.1
pdfminer-six==20221105
tqdm==4.66.1

# API and UI
fastapi==0.109.2
uvicorn==0.25.0
streamlit==1.31.0
pydantic>=2.9.0

# Database
sqlalchemy==2.0.25
chromadb==0.4.22

# Utils
python-dotenv>=1.0.1
networkx==3.2.1
datasets==2.16.1
matplotlib==3.8.2
pandas==2.1.4
pillow==10.2.0
requests==2.31.0

# For calendar integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Testing
pytest==7.4.3

# Desktop application dependencies
PyInstaller==5.10.1
typer==0.9.4
tokenizers>=0.13.2,<0.16.0
tiktoken>=0.7.0
